Organize Playwright API specs by resource and add a writing skill.
